This simple enough refactor seems to have fixed the over-recomposition issue -- I think by passing the noteMatrix explicitly as a MutableState. This means that MIDI notes start *and* stop correctly now.
