+++ app/src/main/java/com/kaelambda/note_grid/MainActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.kaelambda.note_grid.screens.NoteGridScreen
+import com.kaelambda.note_grid.screen.NoteGridScreen
 import com.kaelambda.note_grid.ui.theme.NoteGridTheme
 import dagger.hilt.android.AndroidEntryPoint
