+++ app/src/main/java/com/kaelambda/note_grid/screen/NoteGridScreen.kt
@@ -1,4 +1,4 @@
-package com.kaelambda.note_grid.screens
+package com.kaelambda.note_grid.screen
 
 import androidx.compose.animation.core.Animatable
 import androidx.compose.animation.core.LinearEasing
@@ -11,9 +11,7 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.selection.toggleable
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.Button
 import androidx.compose.material3.Surface
@@ -27,8 +25,6 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -45,7 +41,7 @@ fun NoteGridScreen(viewModel: NoteGridViewModel) {
     var buttonText by remember { mutableStateOf("Play") }
     var randomizationDensity by remember { mutableIntStateOf(25) }
 
-    var noteMatrix by remember {
+    val noteMatrix = remember {
         mutableStateOf(
             Array(xCount) {
                 Array(yCount) { false }
@@ -59,7 +55,7 @@ fun NoteGridScreen(viewModel: NoteGridViewModel) {
             t.animateTo(
                 100f,
                 animationSpec = infiniteRepeatable(
-                    tween(2500, easing = LinearEasing)
+                    tween(5000, easing = LinearEasing)
                 )
             )
         } else {
@@ -69,32 +65,9 @@ fun NoteGridScreen(viewModel: NoteGridViewModel) {
 
     Surface(modifier = Modifier.fillMaxSize()) {
         Column(modifier = Modifier.padding(16.dp)) {
-            for (y in 0 until yCount) {
-                Row {
-                    for (x in 0 until xCount) {
-                        val startTime = remember { duration * x }
-                        val endTime = remember { startTime + duration }
-
-                        val isEnabled = noteMatrix[x][y]
-                        val isPlaying = isEnabled.and(t.value > startTime && t.value < endTime)
-
-                        Note(
-                            y,
-                            isPlaying,
-                            isEnabled,
-                            viewModel::playSound,
-                            viewModel::stopSound
-                        ) {
-                            noteMatrix = noteMatrix.clone().apply {
-                                this[x][y] = it
-                            }
-                        }
-                    }
-                }
-            }
+            NoteGrid(noteMatrix, t, viewModel::playSound, viewModel::stopSound)
 
             Spacer(Modifier.height(16.dp))
-
             Row {
                 Button(onClick = {
                     playing = !playing
@@ -105,14 +78,16 @@ fun NoteGridScreen(viewModel: NoteGridViewModel) {
 
                 Spacer(Modifier.width(16.dp))
                 Button(onClick = {
-                    noteMatrix = Array(xCount) { Array(yCount) { false } }
+                    noteMatrix.value = Array(xCount) { Array(yCount) { false } }
                 }) {
                     Text("Reset")
                 }
 
                 Spacer(Modifier.width(16.dp))
                 Button(onClick = {
-                    noteMatrix = Array(xCount) { Array(yCount) { Random.nextInt(100) < randomizationDensity } }
+                    noteMatrix.value = Array(xCount) {
+                        Array(yCount) { Random.nextInt(100) < randomizationDensity }
+                    }
                 }) {
                     Text("Randomize")
                 }
@@ -130,53 +105,6 @@ fun NoteGridScreen(viewModel: NoteGridViewModel) {
             }
         }
     }
-}
-
-@Composable
-fun Note(
-    noteId: Int,
-    isPlaying: Boolean,
-    isEnabled: Boolean,
-    playSound: (Int) -> Unit,
-    stopSound: (Int) -> Unit,
-    onValueChange: (Boolean) -> Unit
-) {
-    val color = when {
-        isPlaying -> { Color.Green }
-        isEnabled -> { Color.Red }
-        else -> { Color.Gray }
-    }
-
-    if (isPlaying) {
-        playSound(noteId)
-    } else {
-        // This will trigger for notes on the grid that are disabled or not currently playing
-        // but with the same noteId if notes are being recomposed when they shouldn't be
-
-        // Yes -- seems every note is being recomposed when e.g. I enable a single note.
-        // Could this be because isPlaying and isEnabled are vals that are recalculated every
-        // time the screen is recomposed?
-
-        // Perhaps we should hoist the state into a StateHolder for each Note?
-        // Or use remember / mutableStateOf / derivedStateOf in the proper way?
-
-        // When NoteMatrix changes, could that have knockon effects?
-        // A breakpoint here is also hit for every note as t changes while playing
-
-        stopSound(noteId)
-    }
-
-    Surface(
-        Modifier
-            .size(20.dp)
-            .toggleable(
-                value = isEnabled,
-                enabled = true,
-                role = Role.Switch,
-                onValueChange = onValueChange
-            ),
-        color = color
-    ) { }
 }
 
 @Preview(showBackground = true, showSystemUi = true)
